Call addSurface and addCell with data from Parser
Note that I changed the implementation of addCell a bit, so it's doesn't
actually work right now.  I added a FIXME to the implementation of addCell
in Geometry.cpp, but as the code is not line-wrapped I can't actually read
it so I'm not going to try and change it (and end up making a big mess).
I left the implementation of addSurface alone so it's fine.

You'll need to enable c++0x to get this to build, the option
"-std=gnu++0x" should be sufficient on any recent release of G++, I don't
know what should be done for any other compilers.