Change main() to use stack allocation
Using new was unnecessary here, and if you call new then you need to
also call delete (which people will forget) in order to have a clean
valgrind run.  Changing to stack allocation prevents needing to call
delete for a lot of things, and if people forget to call delete then
valgrind won't be clean, which is a pain.