~Parser() deletes all allocated memory.
This means that the parser needs to stay around until after all the
parsed items (surfaces, cells, and lattices) are done being used.  In
the current code this is true (as the only Parser that is created is
allocated on the stack of main()), but now it must remain true.