Detected bounding box automatically
geometry.xml in folder 5 and 6 now have boundary="reflective" in their
surface definitions. surface class got an new attribute called boundary
which is either REFLECTIVE or BOUNDARY_NONE. getXMin, getXMax, getYMin,
getYMax were implemented for each surface class to update the corners of
the bounding box. Inside of addSurface method, every time a new surface
is added, it also updates the corners using the getXMin's methods.
Finally, getHeight and getWidth methods return the differences between
the corners. Tested with 5/geometry.xml, and a bounding box of 4 by 4
is detected and print to screen. Notice to test it the addCell and
addLattice methods have to be commented out in the openmoc.cpp,
otherwise the code errors out before even reaching to the point of
generating bounding box.