Implemented findNextCell method in Geometry.
This method will take in a LocalCoords object and an angle for a
trajectory and determine the next cell that the LocalCoord will find
itself in. This is useful for semgenting tracks since each intersection
point will take place on a surface. From that surface, we want to
know the next cell the LocalCoord will find itself in. This function was
tested with some simple test cases using xml-sample/5/geometry.xml