Edited the README file.
Github displaed an empty README file for our code for now which
looks a bit odd. This README pretty much came from "OpenMOC Coding
Guidelines.txt" (which has been removed, as it's now reduntand)
but there is a little bit of extra information added.
There is an empty space for a list of the mantainers of this
package that needs to be filled in.