Added two methods to Timer class and moved timing solely to openmoc.cpp.
Added a method called recordSplit(const char* msg) to the Timer class.
Call this method with a message after having used the timer to record a
certain process. The Timer will save your message and the elapsed time
in a vector. At the end of the program you can then call the new
printSplits() method and the Timer will print a formatted list of
messages and times to the console.