Completed region index schema.
Method inside Geometry can now take a localcoords linked list and return
the id for the flat source region that it is located in. This method
assumes that the localcoord is 1) within the geometry, 2) that the
linked list is complete down to the lowest level universe, and 3) that
the localcoord passed into the function is at universe 0.