Added number of rings functionality.
Geometry.cpp now supports creating corresponding surfaces and cells
based on the num_rings value for a single circle cell (that is, a
cell described by a single circle surface). Current codes work for
folder 1.1, and prints out the correct number of FSRs. For future,
need to make more test cases and test the track output as well.
Also implementing num_rings for ring object would be the next step.