Complete (nor finalize) number of rings functionality.
Geometry.cpp now supports a number of rings value in the cell definition, for
both circle type cell and ring type cell. Any other type would generate an
error message pointing out whether it is the type of cell that is not supported
(only MATERIAL is suppoprted, not FILL), or the type of surfaces that is not
supported (only 1 circle or 2 circles are supported). This design is not
finalzied because codes need to be cleaned up, and more thoughts need to be put
into generating new index for surfaces/cells.