Fix bugs in num_rings function
OpenMOC runs for all folders now. A previous bug appears from that getNumRings()
is not implemented for FILL type cell, and I called this function to print error
message for user accidentally inputing num_rings for FILL type. Geometry.cpp
would not validate this specific error; Parser would check for it. Also settings
for emacs is updated to make a tab to be 4 spaces instead of 8, so that nested
loops appear a bit easier to read in emacs.