Fixed cutting up cells into rings.
Geometry.cpp and cell.cpp are updated to fix the bugs as shown up
in plots of segments; rings should work for circle cell and ring
cell; ring definitions are added in 3 geometry files (Lulu/1.1,
SimpleLattice, BundleLattice). Future testing is needed to make
sure this functionality works; also a new indexing scheme is
needed for the newly generated surfaces and cells.