Added maps between FSR ids and cells and materials to geometry, and to fluxes in Solver.
Geometry initializes the arrays for FSR maps to cells and materials
inside the constructor. You can access each array with getFSRtoCell and
getFSRtoMaterial methods.

Solver initializes the array for FSR maps to fluxes at the very end of
the computeKeff method. This is a 2D array of fluxes indexed by 1)
energy group and 2) FSR id. You can access each array with getFSRtoFlux
method.