Fixed cutting-up-cell's newly made surfaces and cells' index issue.
The index for newly made surfaces and cells as part of the cutting
cell process are updated to start with 10000 (hence 10000 and above
needs to be reserved for the kernel) and increment as more and more
rings are added. Some sample rings are added in the c5g7 geometry
file as well as test case and they run without failing.