Implemented cutting cut cells into 4 sectors.
Now cells can take both `rings="value"' and `sectors="4"'
definition. Sectors are added in a couple of xml-sample folders,
like the SimpleLattice case and the c5g7 case, and visulazation
confirms that the geometry is indeed cut into the right number
of sectors and rings. This code definitely needs cleanup, and
sector of 8 and 16 should be implemented fairly straight-forward.