Sectors can be added in arbitrary cells.
Previously sectors can only be added in cells with one through
three surfaces. Now sectors can be added into any cells (notice,
cells with surface="" has not been tested). The default case
is changed to PinCell, and it is tested to be working. The
SimpleLattice case generates segmentation fault, which is the
bug Will is trying to fix.