Fixed stray tracks within sectorized cells.
I now move the first localcoords for a track by TINY_MOVE before finding
the cell it is in so as to ensure that it is within the interior of a
cell. I was having problems with tracks that started exactly on a sector
and this led to some stray segments for those tracks.

There is currently an issue with segmenting the simple pin cell.