Implement 8 sectors.
Notice this commit is not fully functioning. I am just pushing it
so the rest of the team can debug. When running a geometry with 8
sectors, currently it would return a error message complaining
about findFSR are looking for FSR in cells with negative or
zero id. This problem only exists for 8 sectors for now.