Tested 16 sectors on BundledLattice
To run Bundlelattice, the default setting is not enoug (will generate
an error message reminding user to put in more angles or more
less spacing), and 0.01 spacing and 32 angles runs, generating
proper plots and k_eff with 16 sectors. Although I cannot visualize
with Visit, so there is a chance the plot is not correct I just cannot
see anything.