Switches to autoconf/automake
This is a pretty major switch, but it's probably for the best.  Despite
autoconf/automake being pretty nasty, they're the standard for portable
open-source software so this will make people not freak out when they
get the software.

I've also created "./autogen.sh", which should manage to clean and
recreate all the autoconf-related files.  You can call it with --clean
to remove all autoconf/automake generated files, or with --distib to
generate files suited for distribution (ie, generate ./configure, check
that it works, and then clean everything up).

Note that autoconf adds a dependency on pkg-config because Magick++
installs its header files in a non-standard location.

Signed-off-by: Lulu Li <lilulu6@gmail.com>