Fixed Mesh::splitCorners in src/Mesh.cpp
Void last commit. In the last commit I was trying to merge the
conflicts in src/Mesh.cpp between Sam and my versions, and in
attempting to do so accidentally use the version of
Mesh::splitCorners that is currently on the master branch. Please
ignore the last commit and this commit should bring us back to
our current Mesh::splitCorners.