This branch's CMFD imeplementation is synced up with CMFD branch.
It turns out the reason the ratio of the cell-averaged flux
is very different from 1.0 at the early 2 iterations is because
my keff convergence threshold is set to be too low. As soon as
you tighten up the convergence to 1e-11, or run many iteration
of MOC before starting CMFD (to get the boundary flux works out)
then it is fine.