MG kind of works!
Minor things are off (for instance, when MOC converges fully then
run LOO with mg somehow the values are changing slightly between
iterations, and results come out to be 0.1 pcm off), but at least
MG converges to the right pcm.
SG on the other hand is giving me a source update factor of 0.05
which does not make any sense.