SG scattering xs is the sum of all groups.
It turns out, if we drive MOC to convergence, and then let LOO
run for one step, with no scattering xs, the source is being
updated to ~0.05 which does not make any sense. Only when we
compute scattering xs by adding up all scattering source in
Cmfd::computeXS(), we can get the source update to be very
close to 1.