Fixed the consistency of $k$ in computing new FSR source.
The point is, we need to use the $k$ that drives this MOC sweep
to compute the new FSR source. If we go ahead and compute the new
$k$ based on the scalar flux of this iteration, then use this $k$
to compute the new FSR source, then the energy group structure in
LOO solver would not be exactly all 1.0 in the debug mode. Also
now we can run two MOC sweeps in LOO now.