Fixed a bug in material's chi normalization routine.
Previously when I normalize chi, I check before doing the division
that the sum of chi is not zero; then I forgot to actually do
anything about chi for the case that chi is zero. That is, water's
chi is just random numbers stored in the memory, which is probably
fine as this happens in the beginning of the problem and the
number stored there is probably zero, but could cause trouble
depends on how lucky I am.