LOO (phi) appears working for one case after using as tracked vol.
That is, when we divide leakage by the side length of the cell,
we should really do as-tracked mesh cell volume divided by real
side length. That is, the surface-integrated leakage is tallied
using real side length, vs the rest of the terms are on as-tracked
volume. This is consistent with the CMFD implementation where the
transport terms (with Dhat and Dtilde) are multipled by real cell
length, and any cross section terms are multiplied by as-tracked
mesh cell volume.