Minor cleanup in Options.cpp; also delete SIGMA_T_THRESH.
Options.cpp is cleanup slightly so that lines are within 80
characters.
SIGMA_T_THRESH was originally used for checking whether adding up
all xs agree with the parsed in total xs as a threshold; now that
we force the total xs to equal to the sum of the absorption xs
and al the scattering xs scattering into this group, there is no
reason for threshold for sigma_t.