Implemented boundary flux update on all four surfaces. pass test.
LOO was not generating the right k before because I was messing up
the quadrature flux index before. For incoming flux on reflective
boundary, we actually store the outgoing flux, thus the index
should be the opposite of the incoming flux. Keep in mind that
all quadrature current and quadrature flux are assumed to be
going out of the cell.