Update boundary flux update routine.
I entirely got rid of the meshCell's _boundary_update and fsr's
_boundary_update mess, because a corner cell can have two
reflective boundaries and thus needs to store four different
numbers. Now I just use the fact that each segment knows which
surface it intersects with, and the new and old quadrature fluxes
are stored on the surfaces. Also check that old quadrature flux
is not zero before dividing.