Update so _diffusion is fully supported in all acceleration.
That is, if _diffusion is on, then both CMFD and LOO would
perform a diffusion acceleration after the very first MOC
transport sweep, and update boundary flux by scalar flux ratio.
More specifically, we only update by quadrature flux when LOO is
on, either _diffusion is off, or if _diffusion is on and we are
not looking at the 1st iteration.