Added in update boundary flux for corner segments for LOOs.
Previously, in updateBoundaryByQuadrature() I forgot to treat
corner segments, because corner segments has no quadrature fluxes
and would not be updated. Currently I update corner segments by
the adjacent surfaces (the same ones that corner segment's flux
was sent to during splitCornerQuadratureCurrent()) updates. This
seems to work well, as now I update every boundary segment
(previously when I check it never quite matches up and I always
wondered why).
This is not an issue for CMFD though, as CMFD uses segments to
find fsr and then to find the mesh; updateBoundaryByHalfSpace()
is blind to whether the segment went through the corner of mesh
cell or not, so CMFD always update all segments.