Minor update of LOO solver's source calculation to debug -bi 0.
I played around with where to put the source calculation in the
iterative LOO solver. When the source calculation is done at the
end of each acceleration (that is, quadrature source is directly
from the high order, instead of being updated by the ratio of
\bar{Q}^{m+1/2} over \bar{Q}^{m} for the first iteration),,
- wl1: does not acceleration;
- wl2: behave the same way as doing the source calculation at
  the beginning of each ierative solve.
So I am back to placing the source calculation in the beginning of
the LOO iterative solve.