Fixed a typo in Cmfd's update boundary flux routine.
When update boundary is requested, I am supposed to save the
boundary flux on each geometry reflective boundary, but I had a
typo there that reads "VACUUM" when it is supposed to be
"REFLECTIVE", so previously it was not saving surface 3's boundary
flux in the forward direction. Upon fixing this bug it looks like
some of the asymmetry situation generated by -ub is reduced.