Implemented functions to determine whether reflectiveness or not.
Given a track in a mesh cell, we can figure out whether it's
going into a reflective/vacuum BC or leaving a reflective/vacuum
BC and do things according with the angular flux (set it to zero
for vacuum, store it for reflective with -ub on). The issue with
LOO fails debug test still exists.