Fix LOO2 balance issue for the new weighting routine.
The weighting routine implemented in commit 335f82032 produces
a slightly different version of the quadrature current that now
the partial current = sum of two quadrature flux divided by 2.
It is easier to visualize it in quadrature flux land, as the
flux representing the half space should be the average of the two
quadrature fluxes in the half space with no weighting, as the two
quadrature fluxes occupy the exact same angular-spacial interval
and thus should be exactly equally weighted!