Implemented real surface-averaged current/flux calculation.
initializeWeights() routine first computes the weights associated
with each quadrature, which is currently approximately the
surface length. It does some quite complicated stuff (e.g., based
on whether a segment hits a verticle or horizontal surface, a
sin or cos of theta_a is divided on the weight, and corner
splitting that's consistent with how the current/flux are
splitted).

Bad news: LOO2 is not quite working yet (for homogeneous
problem partial currents are not the same everywhere). I might
have to remove those cos/sin terms to get homogeneous problem to
work....