Minor update from debugging; ho/lo current calculated faithfully.
The net_current generated from low order is still off compare
with high order. The sum over all geometry is right; the sum
over all spatial region for each energy group is right; looks like
what's going on is the spatial distribution is wrong.

There is also question whether the high order holds balance or
not, which should be tested for the VAC boundary condition case!