Change LOO quadrature weights's corner treatment to = current's.
- How LOO quadrature weights are tallied:
Previously there is a saperate routine Solver::tallyLooWeight(..)
tallies the quadrature weights with built-in corner treatment
that I believe to be consistent with Mesh::splitCornerQuadCurrent.
Evidentally ever since Nov 1 (commit 647f3e12c8f), any case with
vacuum BC would not pass -debug. To be fully consistent, now quad
weights are tallied into all 8 surfaces, then a
Mesh::splitCornerQuadWeights() split the weights the same way the
quad currents are split. This fixes the debug case for cases w/
vacuum BC. Also instead of having 6 _total_wt, now there are only
3, as the first two were never used and _total_wt[2] is changed
to _as_tracked_length as described in the following.

- How as tracked surface lenght is tallied and used:
a Solver::tallyAsTrackedLengthSingle(..) is kept (from previous
algorithm) and called in Solver::tallyLooWeight(..) to calculate
the as tracked length. This length is used in LOO, and appear to
be breaking the case when physical length is used.