Fixed bug in commit d2a0d48769de94643d6de991add8287a55d018d1
commit d2a0d48769de94643d6de991add8287a55d018d1 introduced a
converged flag that is on when LOO converged, and run on more
loo iteration to store the boundary fluxes. Except the first
couple of transport sweeps where the max # of LOO sweeps are
reached, in these case we need to force LOO to store at the
last iteration (max_outer - 1).