A rough estimate cell source (fis + scat) eps for unaccelerated
Basically unaccelerated MOC does not know the mesh cell structure
so there is no clean way to do this. For now I fix the assumption
to be calculating eps on pin cell level, and do a (very bad)
estimate using FSR's source (both fission and scattering) by
examing adjacent FSRs and comparing their value (if the values
are the same then they belong to the same cell etc). This is
definitely not great but works for now!