Test case suite now requests -ub for all acceleration schemes.
LOO1 & LOO2 needs -ub to converge to the right solution. This
make sense because we want -ub to be on to be consistent. I don't
know how LOO1 & LOO2 were working before with -nub (i.e., boundary
angular fluxes were not prolongated). I don't care enough to
figure it out at this point because it probably doesn't help
to keep it working.