Update min_outer for methods. Set CMFD's to 50, LOO's to 80.
This is the minimum number of iterations each method needs to
go through before exiting the low order solver. CMFD's was
previously 5, and changing it to 50 does not seem to change the
convergence behavior for B&W 1810 cores.

For the same B&W 1810 core 1 problem, OpenMOC with LOO seems to
take one more transport sweep when min_outer is reduced to 50 or
30 (so something like 1.2e-5 as opposed to 1e-5 at the 20th
iteration). This is not a huge deal, but for now I am converging
fully by setting min_outer to 80.

Will need to do a sensitivity analysis on min_outer later.