Fixed the messed up bracket introduced in commit 324c28c
In commit 324c28c, when switching from if statement to switch
statement in the Cmfd::computeQuadSrc() routine, the brackets
were messed up, so sumquadflux was not being stored, causing LOO
to diverge on iteration 3.

Also we get some negative numbers early on if src was not forced
to be positive, so they are forced to be positive for now.