Fixed a bug: should only wipe out the counters for one group.
Before, FSR fluxes are zeroed at the beginning of one iteration,
so when one energy group is finished and updateSource() is called,
the FSR fluxes of the lower energy groups are still zero which
generates wrong answers. Current implementation is to only zero
out the FSR fluxes for the energy loop that is going to be looped
over.