Remove setting negative quad src to 0 introduced in commit 0c3a6b2
In commit 0c3a6b2 I thought forcing quad src to go to non-negative
helps with larger quarter core geometry to converge (I need to
re-evaluate those cases). But this breaks the conservation, and
causes LOO to fail the debug mode (as in, run MOC all the way to
convergence, and then enter acceleration).