Added an implementation of the derivative term, comment out assert
The derivative term in the Newton's method can be written in a
different form subsituting two terms with two other terms given
the equity that we are trying to solve for.

I comment out the assertions (e.g., xs is positive, exponential
is larger than 1).

For some reason the regular closure relations (-wlc1) does not
seem to work well with either LOO1 or LOO2 anymore, so I am going
back to commit 09966f to figure out what's going on.