To fix c5g7_cc (8, 0.3) LOO1, set FSR phi to 1e-4 at 1st iter
Changing how FSR phi prolongation (from mesh cell's fluxes to a
small number) seems to fix the issue with LOO1 closure 1 running
c5g7_cc with 8 azimuthal angles and 0.3 track spacing.

I don't like what this does right now (basically first iteration
is wasted with everything set to close to zero?) but I cannot
seem to find a way to get the original thoughts (i.e., setting
FSR fluxes to be cell fluxes, and boundary fluxes to be that /
4pi ) to work well.

Next will look into whether to remove the 1st iteration all
together if all we are going to do is to set everything to zero.