Fix closure #11 first implemented in commit d3d82ab
The uncollided flux tallied in commit d3d82ab is incorrect because
it should only be not counting the source term in the current mesh
cell. That is, when a track crosses a mesh surface, its uncollided
flux counter is set to sync up with the regular flux counter. Then
from there onward inside of this sell, the uncollided flux tally
tallies the flux attenuation and not the source attenuation.