Add closure #12
Closure #12 is very similar to Closure 11 except setting the
forward and backward pair's cross-sections to be their average.
This also attempt to fix the issue that might be associated with
having a vacuum boundary next to a reflective boundary (that the
incoming flux started from the vac boundary would be zero, and
its corresponding outgoing flux is non-zero becaues it picks up
contribution from the reflective boundary).

For C5G7 and B&W type of problems, closure #12 Pretty much perform
the same as closuer 11. Not that interesting.