Important: limit forcing non-negative FSR source to 5 iterations.
Basically there is likely to be negative FSR sources even after
convergence (at least according to the current version). But to
prevent early iterations from diverging, we force the first couple
of iterations' FSR sources to be non-negative. Though the trick is
to not over-do it. For instance, previously I was forcing
non-negative sources during the first 10 iterations, and this
generates a weird bump at iteration 12  in the convergence
behavior of B&W 1810 series core 2, 4, 14, 15, 16, 17, 19 (see
research notes for Dec 3, 2014). At first this bump appeared to
have to do with AIC rods and linear prolongation (the cases where
the bump showed up). But upon furthre investigation it is the
forcing FSR sources to be non-negative during the first 10
iterations. The bump disappeared after the limit was set to 5
iterations.